해쉬태그,포스트태그 기능 1차 완성,유저서비스 1차완성

+++ src/main/java/com/example/blogex/domain/hashtag/dto/HashtagRankDto.java
@@ -9,12 +9,12 @@ import lombok.NoArgsConstructor;
 public class HashtagRankDto {
     private Long id;
     private String tag;
-    private Long cnt;
+    private int cnt;
 
     @Builder
     public HashtagRankDto(Long id,
                           String tag,
-                          Long cnt) {
+                          int cnt) {
         this.id = id;
         this.tag = tag;
         this.cnt = cnt;
